Handling of ships weapons slots
Started work on the weapon slots that will be interchangable during runtime. This will most likely become more generic to be a interchangable comompment.

Got a test version of the correct spirt for the weapons, and a button spawns in the correct location, This is currently done in the generate sprite class but a generate targetable button class is made but not implemented yet but shall create the buttons that will be trigger by the battle Ui class once the ship has been initalised.